requirements.txt
astroid==2.4.2
cachelib==0.1.1   
cachetools==4.1.1 
certifi==2020.6.20
chardet==3.0.4    
click==7.1.2      
colorama==0.4.4   
cycler==0.10.0
Flask==1.1.2
Flask-Session==0.3.2
google-api-core==1.23.0
google-api-python-client==1.12.4
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wordcloud==1.8.0
wrapt==1.12.1